+++ src/main/java/eve/dukeexception/DukeException.java
@@ -1,4 +1,4 @@
-package bob.dukeexception;
+package eve.dukeexception;
 
 public class DukeException extends Exception {
     public DukeException(String s) {
