+++ src/main/java/eve/storage/Storage.java
@@ -1,10 +1,10 @@
-package bob.storage;
+package eve.storage;
 
-import bob.parser.Parser;
+import eve.parser.Parser;
 
-import bob.tasks.Task;
+import eve.tasks.Task;
 
-import bob.tasklist.TaskList;
+import eve.tasklist.TaskList;
 
 import java.io.File;
 import java.io.FileWriter;
