+++ src/main/java/eve/tasks/Event.java
@@ -1,4 +1,4 @@
-package bob.tasks;
+package eve.tasks;
 
 import java.time.LocalDate;
 import java.time.format.DateTimeFormatter;
