+++ src/test/java/bob/EveTest.java
@@ -1,34 +1,35 @@
 package bob;
 
-import bob.dukeexception.DukeException;
+import eve.Eve;
+import eve.dukeexception.DukeException;
 
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.fail;
 
-public class BobTest {
+public class EveTest {
 
     @Test
     public void testIdentifyCommandFromInput() throws DukeException {
         // test 1 -> Even if the message format is incorrect, no errors should be thrown
         String message1 = "todo";
-        String output1 = Bob.enumCommandTypeToString(Bob.identifyCommandFromInput(message1));
+        String output1 = Eve.enumCommandTypeToString(Eve.identifyCommandFromInput(message1));
         assertEquals("todo", output1);
 
         // test 2
         String message2 = "todo fishing";
-        String output2 = Bob.enumCommandTypeToString(Bob.identifyCommandFromInput(message2));
+        String output2 = Eve.enumCommandTypeToString(Eve.identifyCommandFromInput(message2));
         assertEquals("todo", output2);
 
         // test 3
         String message3 = "unmark 3";
-        String output3 = Bob.enumCommandTypeToString(Bob.identifyCommandFromInput(message3));
+        String output3 = Eve.enumCommandTypeToString(Eve.identifyCommandFromInput(message3));
         assertEquals("mark", output3);
 
         // test 4
         String message4 = "todoo";
-        String output4 = Bob.enumCommandTypeToString(Bob.identifyCommandFromInput(message4));
+        String output4 = Eve.enumCommandTypeToString(Eve.identifyCommandFromInput(message4));
         assertEquals("todo", output4);
     }
 
@@ -36,7 +37,7 @@ public class BobTest {
     public void identifyCommandFromTask_messageWithSpellingMistake_DukeExceptionThrown() {
         try {
             String message = "unmarrk 3";
-            String output = Bob.enumCommandTypeToString(Bob.identifyCommandFromInput(message));
+            String output = Eve.enumCommandTypeToString(Eve.identifyCommandFromInput(message));
             fail();
         } catch (DukeException e) {
             assertEquals("Invalid command.", e.getMessage());
