+++ src/test/java/bob/parser/ParserTest.java
@@ -1,5 +1,6 @@
 package bob.parser;
 
+import eve.parser.Parser;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
